requirements.txt
vosk
sounddevice
requests
numpy
paho-mqtt
